Bridging Two Ideal Point Estimates with OLS Mapping

Bridging Two Ideal Point Estimates with OLS Mapping

bridge.olsmap(
  ip1,
  ip2,
  anchorrows.ip1,
  anchorrows.ip2,
  opt = FALSE,
  opt.iter.n = 10000,
  opt.sample.n = 30,
  opt.th.inline = 0.5,
  blend = TRUE,
  blend.th1 = 0.05,
  blend.th2 = 0.15
)

Arguments

ip1
Matrix or data.frame of 'reference' ideal points (i.e., ip2 ideal points will be transformed and mapped to ip1 space). Rows are resepondents and columns are ideal point dimensions. The current code only allows 2 dimensions.

ip2
Matrix or data.frame of ideal points to be transformed.

anchorrows.ip1
Vector of row number of anchoring respondents in ip1.

anchorrows.ip2
Vector of row number of anchoring respondents in ip2. Must be the same length as anchorrows.ip1.

opt
If TRUE, conduct optimization of transformation through RANSAC (random sample consensus). The default is FALSE.

opt.iter.n
Number of iteration in the optimization of transformation matrix.

opt.sample.n
Size of anchoring respondents to be sub-sampled at each iteration of optimization.

opt.th.inline
Upper bound to determine inline respondents at each iteration of optimization. A respondent is considered 'inline' if the distance between transformed ip2 and ip1 goes below this threshold.

blend
If FALSE, do not use blending procedure. The default is TRUE.

blend.th1
If blend==TRUE, first threshold used in 'blending' procedure. If minimum difference between transformed ip2 and ip1 goes below this threshold, the transformed ip2 is replaced with the closest ip1.

blend.th2
If blend==TRUE, second threshold used in 'blending' procedure. If minimum difference between transformed ip2 and ip1 goes below this threshold (but above blend.th1), the transformed ip2 is replaced with the value that intrapolates ip1 that falls within this threshold.

Value

A list with the following elements along with specified argument values:

ip1: Original ip1 matrix.

ip2_trans: Transformed ip2 matrix mapped on ip1 space.

ip2_orig: Original ip2 matrix