Bridging Ideal Point Estimates from Two Data Sets

Toolbox function that Bridges ideal point estimates computed from two separate datasets.

ipbridging(d1, d2, bridge.method = "homography",
  anchors.method = "subsample", anchors.subsample.pr = 0.1,
  anchors.subsample.method = "random", anchors.selectrows.d1 = 1:100,
  anchors.selectrows.d2 = 1:100, anchors.newdata = NULL,
  input.type = "responses", ip.method = "ooc", ip.dims = 2,
  hg.opt.iter.n = 10000, hg.opt.sample.n = 30,
  hg.opt.th.inline = 0.5, hg.blend.th1 = 0.05, hg.blend.th2 = 0.15,
  ...)

Arguments

d1
Matrix or data.frame of 'reference' responses. (i.e., In mapping methods, ideal points based on d2 will be transformed and mapped to ideal point space based on d1). Rows are resepondents and columns are questions/issues. Alternatively, if input.type=="idealpoints", one can directly set matrix or data.frame of ideal points (rows are respondents and columns are ideal point dimensions).

d2
Matrix or data.frame of responses (or ideal points) to be transformed.

bridge.method
Method of bridging. Currently, following methods are aviailable:

"homography" (default): Homography transformation method. Non-parametric method to find optimal transformation matrix to bridge ideal point estimates (calls bridge.homography).

"olsmap": Use OLs regression to map d2 ideal point coordinates on d1 ideal point space. The procedure described in Shor et al. (2010).

"pooling": Pooling method. Simple method to combine d1 and d2 to jointly compute ideal points. The procedure described in Jessee (2016).

anchors.method
Method to select anchors to bridge ideal points. Following values are currently available.

"subsample" (default): Sample subset of d1 and d2 and use them as anchors. Sampled anchors from another dataset are added to the target dataset when computing ideal points on each dataset. See anchors.subsample.pr and anchors.subsample.method for sampling options.

"selectrows": Use this option if specific anchors are already included in both d1 and d2. Row numbers for anchors must be set in anchors.selectrows.d1 and anchors.selectrows.d2, respectively (Two vectors must have the same length and each element must corresponds to the same anchor respondent). This is the only option for anchors selection if input.type=="idealpoints".

"newdata" Use new dataset that only contains anchors. The data.frame assigned in anchors.newdata are appended to both d1 and d2.

Note that "subsample" and "newdata" requires d1 and d2 to have identical columns. This is not required for "selectrows".

anchors.subsample.pr
Proportion of dataset (i.e., d1 and d2) to be sampled as anchor. The propotion is determined by the dataset of smaller size. Used when anchors.method=="subsample" and anchors.subsample.method=="random".

anchors.subsample.method
Method to sample anchors. Following values are currently available:

"random" (default): randomly sample subset of dataset and use them as anchors. The sample size is determined by anchors.subsample.pr.

"selectrows": Use this option if specific anchors are already known ( but not shared between datasets). Row numbers for anchors must be set in anchors.selectrows.d1 and anchors.selectrows.d2, respectively (Two vectors does not have to have the same length).

anchors.selectrows.d1
Must be positive integers. Specify the row numbers of anchors in d1. If anchors.method=="selectrows", must be the same length and with perfect correspondence with anchors.selectrows.d2. If anchors.method=="subsample" and anchors.subsample.method=="selectrows", the above condition is not required.

anchors.selectrows.d2
Must be positive integers. Specify the row numbers of anchors in d2.

anchors.newdata
Matrix or data.frame of anchors. Must have identical columns as d1 and d2. Used if anchors.method=="newdata".

input.type
The input type of d1 and d2. If "responses" (default), dataset values are responses to multiple issues or votes. If "idealpoints", dataset values are ideal points.

ip.method
Method of ideal point computation (used only when input.type=="response". Following values are currently available.

"ooc" (default): Ordered optimal classification. Calls oocflex function.

"oc": Optimal classification. Calls oc function from oc package.

"wnominate": W-NOMINATE. Calls wnominate function from wnominate package.

"irtMCMC": IRT model via Markov chain Monte Carlo method. Calls ideal function from pscl package.

For ip.method=="ooc", the inputs d1 and d2 must be positive integer. For all other methods, the inputs must be a dummy variable indication 0=Nay and 1=Yea (and 9 for unit nonresponse). All methods allow item non-response as NA.

ip.dims
Number of dimension in ideal point computation. Must be a positive integer between 1 and 10. If bridge.method=="homography", only 2 is accepted at this point.

hg.opt.iter.n
Used if bridge.method=="homography". Number of iteration in the optimization of transformation matrix.

hg.opt.sample.n
Used if bridge.method=="homography". Size of anchoring respondents to be sub-sampled at each iteration of optimization.

hg.opt.th.inline
Used if bridge.method=="homography". Upper bound to determine inline respondents at each iteration of optimization. A respondent is considered 'inline' if the distance between transformed ip2 and ip1 goes below this threshold.

hg.blend.th1
Used if bridge.method=="homography". First threshold used in 'blending' procedure. If minimum difference between transformed ip2 and ip1 goes below this threshold, the transformed ip2 is replaced with the closest ip1.

hg.blend.th2
Used if bridge.method=="homography". Second threshold used in 'blending' procedure. If minimum difference between transformed ip2 and ip1 goes below this threshold (but above blend.th1), the transformed ip2 is replaced with the value that intrapolates ip1 that falls within this threshold.

...
Additional arguments passed to the ideal point estimation function called in ip.method.

Value

A list with the following elements along with specified argument values

ip1: Original ip1 matrix.

ip2_trans: Transformed ip2 matrix mapped on ip1 space.

ip2_orig: Original ip2 matrix

References

Jessee, S. (2016), '(How) Can We Estimate the Ideology of Citizens and Political Elites on the Same Scale?', American Journal of Political Science 60(4), 1108--1124.

Shor, B.; Berry, C. & McCarty, N. (2010), 'A Bridge to Somewhere: Mapping State and Congressional Ideology on a Cross-institutional Common Space', Legislative Studies Quarterly 35(3), 417--448.

See also