Estimating Ideal Points

Toolbox function that estimates ideal points using different methods.

ipest(
  d,
  method = "ooc",
  dims = 2,
  polarity = c(1, 1),
  irt.type = "binary",
  ...
)

Arguments

d
A matrix with numeric values. With following conditions:

Missing values (if any) must be set to NA.

If variables are binary, values should be conisted from 0 and 1.

If variables are ordered, values should be positive integers starting from 1.

method
Method of ideal point computation (used only when input.type=="response". Following values are currently available.

"ooc" (default): Ordered optimal classification. Calls oocflex function.

"blackbox": Blackbox scaling. Calls blackbox function from blackbox package (minscale=10 by default).

"oc": Optimal classification. Calls oc function from oc package (polarity=c(1,1) by default).

"wnominate": W-NOMINATE. Calls wnominate function from wnominate package (polarity=c(1,1) by default).

"irtMCMC": IRT model via Markov chain Monte Carlo method. Calls ideal function from pscl package if irt.type=="binary"; Calls MCMCfactanal function from MCMCpack package if irt.type=="ordered" (still provisional).

"irtEM": IRT model via EM algorithm method (only one dimensional ideal points. This option is still provisional). Calls binIRT function from emIRT package if irt.type=="binary"; Calls ordIRT function from emIRT package if irt.type=="ordered" (Only 3 category variable can be used).

For method=="ooc", the inputs d1 and d2 must be positive integer. For all other methods, the inputs must be a dummy variable indication 0=Nay and 1=Yea (and 9 for unit nonresponse). All methods allow item non-response as NA.

dims
Number of dimension in ideal point computation. Must be a positive integer between 1 and 10. If bridge.method=="homography", only 2 is accepted at this point.

polarity
A vector specifying the row number of the d1(or pooled data)'s respondent(s) constrained to have a positive (i.e., right-wing or conservative) score on each dimension. Used when method is "ooc", "oc", or "wnominate".

irt.type
Used if method is "irtMCMC" or "irtEM". If "binary" the binary IRT will be estimated. If "ordered", the ordered IRT will be estimated (the ordered option is provisional).

...
Additional arguments passed to the ideal point estimation function called in method.

Value

A list with the following elements

ip: Matrix of idealpoints.

info: Outputs including other information.

References

Jessee, S. (2016), '(How) Can We Estimate the Ideology of Citizens and Political Elites on the Same Scale?', American Journal of Political Science 60(4), 1108--1124.

Shor, B.; Berry, C. & McCarty, N. (2010), 'A Bridge to Somewhere: Mapping State and Congressional Ideology on a Cross-institutional Common Space', Legislative Studies Quarterly 35(3), 417--448.

See also