Ordered Optimal Classification (Flexible Version)

Performs Ordered Optimal Classification (OOC) with flexible strategies of estimation. OOC is an extension of Poole's (2000) nonparametric unfolding procedure for the analysis of ordinal choice data (e.g., “Strongly Agree,” “Somewhat Agree,” “Somewhat Disagree,” “Strongly Disagree”).

oocflex(
  votemat,
  dims = 2,
  minvotes = 10,
  lop = 0.001,
  polarity = c(1, 1),
  iter = 25,
  nv.method = "svm.reg",
  binary.method = "dominance",
  random.seed = NULL,
  ...
)

Arguments

votemat
A matrix of ordinal choice data, must be consecutive integers starting with 1. Binary choice should be coded so that 1 = Support, 2 = Oppose. Missing data must be coded as NA.

dims
Number of dimensions to estimate.

minvotes
Minimum number of votes required for a respondent to be included in the analysis.

lop
A proportion between 0 and 1, the cut-off used for excluding lopsided votes.

polarity
A vector specifying the row number of the respondent(s) constrained to have a positive (i.e., right-wing or conservative) score on each dimension.

iter
Number of iterations of the modified Optimal Classification algorithm.

nv.method
The method used to compute normal vectors at each step of the iteration. Current choices include: "oprobit" (Ordered probit regression), "ologit" (Ordered logistic regression), "svm.reg" (SVM: regression), "svm.class" (SVM: classification), and "krls" (Kernel regularized least squares). One can also assign a custom-made function to compute normal vectors. See more details in compute.nv.

binary.method
The method used to transform ordinal responses to binary ones. Following options are currently available:

"dominance" (default): Transformed to the matrix with binary choices organized in a dominance pattern.

"even": Transformed to the matrix with binary choices recoded to be as evenly-balanced as possible.

random.seed
If not NULL, scalar indicates the random seed for reproducibility.

...
Additional arguments passed to the function assigned by nv.method (See more details in compute.nv).

Value

An oocflexObject with the following elements

respondents: A matrix containing the respondent estimates.

issues: A matrix containing the estimates for the (c-1) categories of each issue.

issues.unique: A matrix containing only the estimates for the unique issues.

fits: The percentage of all binary choices correctly predicted, the Aggregate Proportional Reduction in Error (APRE) statistic for the binary choices, the percentage of all ordinal choices correctly predicted, and the Aggregate Proportional Reduction in Error (APRE) statistic for the ordinal choices.

OC.result.binary: Standard (binary) OC results for the choice matrix arranged in a binary dominance pattern.

errorcount: Error counts at each iteration of the Optimal Classification algorithm.

votemat.binary: The matrix of ordinal choices recoded into binary categories that are as balanced as possible.

References

Hare, C., Liu, T., & Lupton, R. N. 2018. "What Ordered Optimal Classification Reveals about Ideological Structure, Cleavages, and Polarization in the American Mass Public", Public Choice, 176, 57-78.

Armstrong, David A., Ryan Bakker, Royce Carroll, Christopher Hare, Keith T. Poole, and Howard Rosenthal. 2014. Analyzing Spatial Models of Choice and Judgment with R. Boca Raton, FL: CRC Press.

Poole, Keith T. 2000. "Nonparametric unfolding of BInary Choice Data." Political Analysis 8(e): 211-237.

See also