Ordered Optimal Classification Summary

summary.oocflexObject reads an Ordered Optimal Classification object exported by oocflex function and prints a summary.

# S3 method for oocflexObject
summary(object, verbose = FALSE, ...)

Arguments

object
an oocflexObject output object.

verbose
logical, includes all ideal points if TRUE, otherwise only returns the first 10 respondents.

...
Other arguments. Do nothing and are not passed to any functions.

Value

A summary of a oocflexObject.

See also