Generate Simulated Responses of Respondents with Ideal Points

Generate simulated responses of survey respondents with ideal points through Monte Carlo method.

genip.montecarlo(n = 1200, q = 20, ncat = 5, ndim = 2,
  utility.probs = c(0.33, 0.33, 0.33), missing = 0.1,
  correlations.lim = c(-0.1, 0.7), error.respondents = c(0.1, 1),
  error.issues = c(2, 1), idealpoints.lim = c(-2, 2))

Arguments

n
Number of respondents.

q
Number of issues.

ncat
Number of ordered category in responses.

ndim
Number of dimentions in ideal points.

utility.probs
Sampling weights for the utility function form. Must be a numeric vector with length 3. The three weights corresponds to linear, normal, and qudratic utility functions, respectively.

missing
Proportion of missing responses. Ranges from 0-1.

correlations.lim
Limits in the range of correlation between ideal points dimensions.

error.respondents
The lower and upper limits of uniform distribution that are used to initiate values of respondents.

error.issues
The shape and rate parameter in rgamma distribution, respectively. Used to set initial values of issue positions.

idealpoints.lim
Limits in the range of ideal points. The values outside of this range is recoded into minimum and maximum value of the range.

Value

A list with the following elements

simulated.responses: n*q data.frame of simulated responses.

perfect.responses: Hypothetical responses if each respondents give responses without error.

idealpoints: Ideal point coordinates of each respondent.

normalvectors: Normal vectors.

heteroskedastic.respondents: Initial values assigned to respondents.

heteroskedastic.issues: Initial values assigned to issues.

correlations: Correlation between ideal point dimensions.

knowledge: Correlation binned into three categories.

error: Proportions of incorrect choices.