Estimating Normal Vector through Support Vector Machine

Compute normal vector from OC coordinates and ordered responses.

nv.svm(
  xmat,
  resp,
  kernel = "radial",
  tune.param = "heuristics",
  use.gpu = FALSE,
  param.heuristics.k = 3,
  param.heuristics.frac = 0.5,
  ...
)

Arguments

xmat
Matrix of OC coordinates (i.e., predictors).

resp
Response Variable (i.e., ordered choices).

kernel
The kernel used in training and predicting. The default is "radial" to use radial basis function (RBF). The alterantaives are "linear", "polynomial" or "sigmoid". See the help file of svm for more details.

tune.param
Method to determine the parameters. Following options are currently available:

NULL: Use svm with the default or manually set parameter values. To manually set parameters. Check the available parameters (and their default values) in svm function in e1071 package.

"heuristics" (default): If kernel=="radial", Use heuristics-based method (i.e., formulation identical to sigest in kernlab package) to determine optimal gamma. Additionally, if resp is numeric, use heuristic based method described in Cherkassky & Ma (2004) to determine optimal parameters for C (cost) and epsilon (epsilon-regression is automatically selected). If resp is a factor, the same behavior as NULL.

A named list of parameter vectors spanning the sampling space. If list is assigned here, the method use best.tune function to obtain the best perfomed model in the given parameter ranges.

use.gpu
(THIS OPTION IS BETA VERSION, NOT CONFIRMED TO WORK) If TRUE, the method replaces best.tune and svm functions with the best.tune and svm functions in Rgtsvm package, which contains GPU accelerate method of svm function. Install Rgtsvm manually. You need to have Linux computer with NVIDIA GPU for the installation. Check https://github.com/Danko-Lab/Rgtsvm for more details.

param.heuristics.k
If tune.param=="heuristics", the method uses k nearest neighbors regression to estimate the noise variance in response variable for the detemination of epsilon. This argument sets k for this method. The default is 3. Cherkassky & Ma (2004) recommends somewhere between 2 to 6.

param.heuristics.frac
Fraction of data to use for heuristics-based estimation of gamma. By default, 50 percent of the data is used to estimate the range of the gamma hyperparameter.

...
Additional arguments passed to svm or best.tune.

Value

A vector of coefficients.

References

Chang, C. & C. Lin, LIBSVM: A Library for Support Vector Machines, http://www.csie.ntu.edu.tw/~cjlin/libsvm.

Cherkassky, V. & Y. Ma, 2004, "Practical Selection of SVM Parameters and Noise Estimation for SVM Regression", Neural Networks, 17, 113 - 126.

Cotter, A., N. Srebro, & J. Kreshet, 2011, "A GPU-Tailored Approach for Training Kernelized SVMs", 17th ACM SIGKDD Conference on Knowledge Discovery and Data Mining.

See also