Estimating Normal Vector through Support Vector Machine

Estimating Normal Vector through Support Vector Machine

nv.svm(xmat, resp, tune.param = "heuristics", tune.gpu = FALSE,
  param.heuristics.k = 3, ...)

Arguments

xmat
Matrix of OC coordinates (i.e., predictors).

resp
Response Variable (i.e., ordered choices).

tune.param
Method to determine the parameters. Following options are currently available:

NULL: Use svm with the default or manually set parameter values. To manually set parameters. Check the available parameters (and their default values) in svm function in e1071 package.

"heuristics" (default): If resp is numeric, use heuristic based method described in Cherkassky & Ma (2004) to determine optimal parameters for C (cost) and epsilon. If resp is a factor, the same behavior as NULL.

A named list of parameter vectors spanning the sampling space. If list is assigned here, the method use best.tune function to obtain the best perfomed model in the given parameter ranges.

tune.gpu
If TRUE and tuning is conducted, the method replaces best.tune function with the best.tune function in Rgtsvm package, which contains GPU accelerate method of svm function. Check https://github.com/Danko-Lab/Rgtsvm for more details.

param.heuristics.k
If tune.param=="heuristics", the method uses k nearest neighbors regression to estimate the noise variance in response variable. This argument sets k for this method. The default is 3. Cherkassky & Ma (2004) recommends somewhere between 2 to 6.

...
Additional arguments passed to svm or best.tune.

Value

A vector of coefficients.

References

Chang, C. & C. Lin, LIBSVM: A Library for Support Vector Machines, http://www.csie.ntu.edu.tw/~cjlin/libsvm.

Cherkassky, V. & Y. Ma, 2004, "Practical Selection of SVM Parameters and Noise Estimation for SVM Regression", Neural Networks, 17, 113 - 126.

Cotter, A., N. Srebro, & J. Kreshet, 2011, "A GPU-Tailored Approach for Training Kernelized SVMs", 17th ACM SIGKDD Conference on Knowledge Discovery and Data Mining.

See also